api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
PyYAML==6.0.2
requests